Add SetHook and SetPanic.
SetHook installs a synchronous callback that is invoked when Run is about to
terminate the program. It cannot prevent termination, but since termination
stalls until it returns, it can be used to clean up resources.

SetPanic controls whether exit occurs by calling os.Exit, the default, or by
propagating a panic. This is mainly useful for testing, but can also be used to
run a subordinate plugin without terminating the wrapper.